+++ utils/vec3.cpp
@@ -8,14 +8,14 @@ vec3::vec3()
 	x = y = z = 0;
 }
 
-vec3::vec3(float x, float y, float z) 
+vec3::vec3(double x, double y, double z) 
 {
 	this->x = x;
 	this->y = y;
 	this->z = z;
 }
 
-vec3& vec3::set(float x, float y, float z) {
+vec3& vec3::set(double x, double y, double z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
@@ -45,12 +45,12 @@ vec3 vec3::operator-(const vec3& v) const {
 	return vec3(x - v.x, y - v.y, z - v.z);
 }
 
-vec3 vec3::operator*(float scalar) const {
+vec3 vec3::operator*(double scalar) const {
 	return vec3(x * scalar, y * scalar, z * scalar);
 }
 
-vec3 vec3::operator/(float scalar) const {
-	float inv = 1.0f / scalar;
+vec3 vec3::operator/(double scalar) const {
+	double inv = 1.0f / scalar;
 	return vec3(x * inv, y * inv, z * inv);
 }
 
@@ -75,32 +75,32 @@ vec3& vec3::operator-=(const vec3& v) {
 	return *this;
 }
 
-vec3& vec3::operator*=(float scalar) {
+vec3& vec3::operator*=(double scalar) {
 	x *= scalar;
 	y *= scalar;
 	z *= scalar;
 	return *this;
 }
 
-vec3& vec3::operator/=(float scalar) {
-	float inv = 1.0f / scalar;
+vec3& vec3::operator/=(double scalar) {
+	double inv = 1.0f / scalar;
 	x *= inv;
 	y *= inv;
 	z *= inv;
 	return *this;
 }
 
-float& vec3::operator[](int index) {
+double& vec3::operator[](int index) {
 	assert(index >= 0 && index < 3);
 	return (&x)[index];
 }
 
-const float& vec3::operator[](int index) const {
+const double& vec3::operator[](int index) const {
 	assert(index >= 0 && index < 3);
 	return (&x)[index];
 }
 
-float vec3::dot(const vec3& v) const {
+double vec3::dot(const vec3& v) const {
 	return x * v.x + y * v.y + z * v.z;
 }
 
@@ -112,18 +112,18 @@ vec3 vec3::cross(const vec3& v) const {
 	return result;
 }
 
-float vec3::magnitude(void) const {
+double vec3::magnitude(void) const {
 	return sqrtf(x * x + y * y + z * z);
 }
 
 vec3& vec3::normalize(void) {
-	float mag = sqrtf(x * x + y * y + z * z);
+	double mag = sqrtf(x * x + y * y + z * z);
 
 	if (mag < 1e-6f) {
 		x = y = z = 0;
 	}
 	else {
-		float inv = 1.0f / mag;
+		double inv = 1.0f / mag;
 
 		x *= inv;
 		y *= inv;
@@ -133,15 +133,15 @@ vec3& vec3::normalize(void) {
 	return *this;
 }
 
-float* vec3::ptr(void) {
+double* vec3::ptr(void) {
 	return &x;
 }
 
-const float* vec3::ptr(void) const {
+const double* vec3::ptr(void) const {
 	return &x;
 }
 
-vec3 operator*(float scalar, const vec3& v) {
+vec3 operator*(double scalar, const vec3& v) {
     printf("x: %f y: %f z: %f\n", v.x * scalar, v.y * scalar, v.z * scalar);
 	return vec3(v.x * scalar, v.y * scalar, v.z * scalar);
 }
